Reorganize b9
move includes to a single directory. Put all includes under a b9
directory as well.

Remove b9 from most source file names.

Rename base9 to b9run.

Move each executable or library's sources into a unique subdirectory.

Move docs to doc/

Fix up the readme with cmake instructions
