+++ .gitignore
@@ -1,6 +1,5 @@
 ###  B9 ###
 
-b9
 omr
 node_modules
 build
