+++ b9test/b9test.cpp
@@ -1,4 +1,4 @@
-#include "base9.hpp"
+#include <b9.hpp>
 
 #include <stdio.h>
 #include <stdlib.h>
